+++ PhotoGallery.Application/Abstractions/IPhotoStorageService.cs
@@ -1,5 +1,8 @@
 ﻿namespace PhotoGallery.Application.Abstractions
 {
+    /// <summary>
+    /// Strategy pattern ( Open/Closed Principle and DI principle )
+    /// </summary>
     public interface IPhotoStorageService
     {
         Task<string> SaveAsync(Stream stream, string fileName, string contentType, CancellationToken cancellationToken = default);